Added policies to comments and posts

+++ app/Http/Controllers/CommentController.php
@@ -10,6 +10,11 @@ use Illuminate\Support\Facades\DB;
 class CommentController extends Controller
 {
     public function store(Request $request, $postId){
+
+        if ($request->user()->cannot('store', Comment::class)) {
+            return response()->json(['message' => 'You do not have permission to create this resource'],403);
+        }
+
         $validatedData = $request->validate([
             'content' => 'required|string',
         ]);
@@ -20,22 +25,26 @@ class CommentController extends Controller
 
         DB::transaction(function () use ($post, $validatedData, &$comment) {
             $comment = $post->comments()->create([
-                'author_id' => 1,               // TODO: Replace with AUTH(user)
+                'author_id' => auth()->id(),
+                'post_id' => $post->id,
                 'content' => $validatedData['content'],
             ]);
             $post->increment('comment_count');
         });
 
-        return response()->json($comment, 201);
+        return response()->json(['message' => 'Successfully left a comment.'], 201);
     }
 
-    public function destroy($postId, $commentId){
+    public function destroy(Request $request, $commentId){
+
+        if ($request->user()->cannot('destroy', Comment::class)) {
+            return response()->json(['message' => 'You do not have permission to delete this resource'],403);
+        }
+
         $comment = Comment::where('id', $commentId)
-            ->where('post_id', $postId)
-            ->where('author_id', 1)  // TODO: Replace with AUTH(user)
             ->firstOrFail();
 
-        $post = Post::findOrFail($postId);  // look for post id
+        $post = Post::findOrFail($comment->post_id);  // look for post id
 
         DB::transaction(function () use ($post, $comment) {
             $comment->delete();
@@ -45,7 +54,12 @@ class CommentController extends Controller
         return response()->json(null, 204);
     }
 
-    public function index($postId){
+    public function index(Request $request, $postId){
+
+        if ($request->user()->cannot('index', Comment::class)) {
+            return response()->json(['message' => 'You do not have permission to view this resource'],403);
+        }
+
         $comments = Comment::where('post_id', $postId)->get();
 
         return response()->json($comments, 201);
